\name{soilwaterptf-package}
\alias{soilwaterptf-package}
\alias{soilwaterptf}
\docType{package}
\title{Wosten et al. 1999 pedotransfer functions for soil hydraulic properties.}
\description{Pedotransfer functions (PTFs) for estimating the physical parameters of soil water retention and hydraulic conductivity functions from easily available soil properties (typically soil texture, organic matter content, bulk density). The pedotransfer function(s) implemeneted is (are): (1) Wosten et al. 1999 (Geoderma 90:169-185) pedotransfer functions for estimating the parameters of van-Genuchten (1980) Mualem (1976) water retention and hydraulic conductivity functions (valid for European soils).}
\details{
\tabular{ll}{Package: \tab soilwaterptf\cr
Version: \tab 1.1.1\cr
Date: \tab 2014-10-02\cr
Title: \tab Wosten et al. 1999 pedotransfer functions for soil hydraulic properties.\cr
Author: \tab Julien Moeys [aut, cre], Till Francke [ctb] <francke@uni-potsdam.de>\cr
Authors@R: \tab c( person( "Julien", "Moeys", role = c("aut", "cre"), email = "jules_m78-soiltexture@yahoo.fr" ), person( "Till",   "Francke", role = "ctb", email = "francke@uni-potsdam.de" ) )\cr
Maintainer: \tab Julien Moeys <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 3.1.1), soilwaterfun\cr
License: \tab AGPL-3\cr
URL: \tab http://soilwater.r-forge.r-project.org/\cr
Imports: \tab soiltexture\cr}
}
\author{Julien Moeys [aut, cre], Till Francke [ctb] <francke@uni-potsdam.de>}

\keyword{ package }


