\name{soilwaterptf-package}
\alias{soilwaterptf-package}
\alias{soilwaterptf}
\docType{package}
\title{
\packageTitle{soilwaterptf}
}
\description{
\packageDescription{soilwaterptf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{soilwaterptf}
\packageIndices{soilwaterptf}
}
\author{
\packageAuthor{soilwaterptf}

Maintainer: \packageMaintainer{soilwaterptf}
}
\keyword{ package }
