\name{soilwaterptf-package}
\alias{soilwaterptf-package}
\docType{package}
\title{Wosten et al. 1999 pedotransfer functions for soil hydraulic properties.}
\description{Pedotransfer functions (PTFs) for estimating the physical parameters of soil
water retention and hydraulic conductivity functions from easily available soil
properties (typically soil texture, organic matter content, bulk density). The
pedotransfer function(s) implemeneted is (are): (1) Wosten et al. 1999 (Geoderma
90:169-185) pedotransfer functions for estimating the parameters of van-Genuchten (1980)
Mualem (1976) water retention and hydraulic conductivity functions (valid for European
soils).}
\details{
\tabular{ll}{Package: \tab soilwaterptf\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-19\cr
Title: \tab Wosten et al. 1999 pedotransfer functions for soil hydraulic properties.\cr
Author: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
Maintainer: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.11.1), MASS\cr
Suggests: \tab soilwaterfun\cr
License: \tab GPL-3\cr
URL: \tab http://www.mark.slu.se, http://ckb.slu.se, http://julienmoeys.free.fr\cr}
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}

\keyword{ package }


